+++ CHANGELOG.md
@@ -1,15 +1,21 @@
 # Changelog
 
-# 1.0.2
+## 1.1.0
+
+* added reading client_ip from Caddy to allow the usage of proxies
+* allow configuring a client secret instead of an oAuth client
+* updated dependencies
+
+## 1.0.2
 
 * fixed dependencies
 
-# 1.0.1
+## 1.0.1
 
 * fixed go.mod
 * updated dependencies
 
-# 1.0.0
+## 1.0.0
 
 * forked repository
 * added vendoring